Doing the disaggregation in terms of rates instead of probabilities
